requirements.txt
prometheus_client
requests
pyyaml